moved random color
