+++ src/app/pages/feed/feed-overview-page/feed-overview-page.component.ts
@@ -3,7 +3,7 @@ import { Apollo } from 'apollo-angular';
 import { GENERAL_FEED_QUERY, GENERAL_FEED_QUERY_RESPONCE } from '../../../core/queries/feeds';
 import { GENERAL_SUBSCRIPTIONS_FROM_USER_QUERY } from 'src/app/core/queries/subscriotionsFromUser';
 import { GENERAL_FEED_SUB_MUTATION } from 'src/app/core/mutations/subsscription';
-import { DEFAULT_PASTEL_COLOR, generateRandomColor } from 'src/app/core/utils/randomColor';
+import { DEFAULT_PASTEL_COLOR, generateRandomColor } from 'src/app/util/randomColor';
 import { ActivatedRoute } from '@angular/router';
 import { KeycloakService } from 'keycloak-angular';
